Implement current user id parameter validation for controller methods
Validation is implemented using Spring AOP. This way has been chosen
for possibility to make method-level validation and have wider
capabilities to extend validation functionality in the future.
The validated parameter has to be of type Long and be annotated with
CurrentUserId annotation. Also, the parameter can be placed at
an arbitrary position in the method signature.
Currently, only greencity.controller package is handled because it is
the only place with such a validation logic. However, package scope can
be extended or changed completely by rewriting pointcut expression in
the aspect class.
See this aspect usage in the future commits.