Initial implementation of RabbitMQ messaging for EmailService
This commit introduces a RabbitMQ messaging config which defines
key exchanges and queues for sending password recovery emails.
Also, the message receiver is implemented for those queues.
Currently, all the classes are defined in the current repository, but
some of them will be moved to separate microservice soon, once
all the rest email-related application events are switched for
using the RabbitMQ messaging system.