Fixed Bug
Bug was fixed but the problem is deeper. In controller exist operation with endpoint that starts with /{userId}/all-users-friends and when we put null in delete(/friends/) Spring security cannot understand which endpoint it should choose Get or Delete, so we have a 403 error instead of 400. I fixed this problem by adding additional Mapping, but this problem should be revised deeper